feat(mobile): unify volunteer profile screen and add taps everywhere users appear
Merges /friend/[id] and /user/[id] into a single /user/[id] route that
renders the trimmed (not-friends) or full (friends) view from one fetch.
The /api/mobile/users/[id] endpoint now returns a `connection` block
(shared shifts, monthly rhythm, favourite role, timeline, target's
upcoming shifts) only when the viewer is an accepted friend.

Wires up tappable profile links across the app: shift roster rows,
feed-card avatars (achievement, milestone, photo_post, friend_signup),
likers stack in the feed sheet, and the achievement-sheet "friends who
earned" list.
