+++ mobile/app/_layout.tsx
@@ -147,15 +147,6 @@ export default function RootLayout() {
             }}
           />
           <Stack.Screen
-            name="friend/[id]"
-            options={{
-              headerShown: true,
-              headerTransparent: true,
-              headerBackTitle: 'Back',
-              title: '',
-            }}
-          />
-          <Stack.Screen
             name="user/[id]"
             options={{
               headerShown: true,
